FIX MONSTER SO IT DOESN'T CHANGE WHEN ATTACKING
I user React.memo for it, which basically wraps a child component and only updates it when its
props update, but not if the parent updates, which is what was happening. It still happens for
the Hero, but it doesn't affect its appearance and I am not sure React.memo would work, because
Hero now has state.

I also commented out a sound that is not used and the Spacebar click, which did not actually
do anyting, apart from causing the component to re-render unecessarily. You can still use the
Spacebar after you start the attack by clicking on the button. So it was not the function that
was doing it. I left it in commented out, if anyone has any doubts.